UCMAIAgent/myproject/pyproject.toml
[project]
name = "chatgpt-clone"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "graphviz>=0.21",
    "openai-agents[viz]>=0.2.6",
    "python-dotenv>=1.1.1",
    "streamlit>=1.48.0",
]

[dependency-groups]
dev = [
    "ipykernel>=6.30.1",
]

[project.scripts]
ucm-agent = "main:main"

[tool.setuptools]
py-modules = ["main"]

